Models update and fix of controllers after update

+++ CourierAppBackend/Models/Price.cs
@@ -6,8 +6,8 @@ namespace CourierAppBackend.Models;
 [Owned]
 public class Price
 {
-    public float FullPrice { get; set; }
-    public float Taxes { get; set; }
-    public float Fees { get; set; }
-    public float Value { get; set; }
+    public decimal FullPrice { get; set; }
+    public decimal Taxes { get; set; }
+    public decimal Fees { get; set; }
+    public decimal Value { get; set; }
 }
